feat: add handling of nonZero type
